% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_times.R
\name{fun_times}
\alias{fun_times}
\title{Function that provides your age}
\usage{
fun_times(your_age)
}
\arguments{
\item{your_age}{A parameter for your current age}
}
\value{

}
\description{
Function that provides your age
}
